Big MGS3 Update Minor MGS2 Update
Added about 100 effects to MGS3! Should be good for a substantial test, should be good for a public release on CC's platform after testing. On MGS2 I fixed some issues I had with effects permanently failing because of an oversight on my part, need to add a lot more effects to MGS2 before a good public release for it is viable.